add homework module06

+++ JavaOnline_Homework/src/ua/goit/gojavaonline/core_3/groupe_02/chastnikov/module04/temperature/TemperatureConverter.java
@@ -2,11 +2,9 @@ package ua.goit.gojavaonline.core_3.groupe_02.chastnikov.module04.temperature;
 
 public class TemperatureConverter {
 
-    static float toCelsius(float tempFahrenheit) {
-        return (tempFahrenheit - 32) * 5/9;
-    }
+    static double toCelsius(double tempFahrenheit) { return (tempFahrenheit - 32) * 5/9; }
 
-    static float toFahrenheit(float tempCelsius) {
+    static double toFahrenheit(double tempCelsius) {
         return tempCelsius * 9/5 + 32;
     }
 }
